# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quiz App backend API",
        "description": "Documentação da API desenvolvida para o quiz app",
        "contact": {
            "email": "marcospastora@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/app/login/professor": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "login - api.app.login.professor",
                "description": "Login professor",
                "operationId": "62a5b13d1eed936ad448494456549b7e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "professor@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "abc123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/app/login/aluno": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "login - api.app.login.aluno",
                "description": "Login aluno",
                "operationId": "2b042dcf3552033b7d6895ec6f5225c9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "aluno@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "abc123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/quiz/status/listar": {
            "get": {
                "tags": [
                    "Quiz"
                ],
                "summary": "quiz - api.quiz.status.listar",
                "description": "Listagem status do quiz",
                "operationId": "295ed253e1f4b5780214eed2af446404",
                "responses": {
                    "200": {
                        "description": "Successful"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/app/usuarios/cadastro-professor": {
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Cadastro - api.app.usuarios.cadastro.professor",
                "description": "Cadastro de usuário do tipo professor",
                "operationId": "e5697229d8098d7ce18f978c418f1db5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "nome",
                                    "registro_professor",
                                    "data_nascimento",
                                    "whatsapp"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "professor@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "abc123"
                                    },
                                    "nome": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Professor Girafales"
                                    },
                                    "registro_professor": {
                                        "type": "string",
                                        "format": "number",
                                        "example": "558741"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1996-09-28"
                                    },
                                    "whatsapp": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "27996123123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/app/usuarios/cadastro-aluno": {
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Cadastro - api.app.usuarios.cadastro.aluno",
                "description": "Cadastro de usuário do tipo aluno",
                "operationId": "048500a0a77df84e3388429b9d01382b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "nome",
                                    "matricula",
                                    "data_nascimento",
                                    "whatsapp"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "aluno@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "abc123"
                                    },
                                    "nome": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Chaves Del Oitcho"
                                    },
                                    "matricula": {
                                        "type": "string",
                                        "format": "number",
                                        "example": "558741"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1996-09-28"
                                    },
                                    "whatsapp": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "27996123123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "realize login para obter o token",
                "name": "Token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}